requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
jupyter>=1.0.0
notebook>=7.0.0
python-dotenv>=1.0.0
pyarrow>=14.0.1
fastparquet>=2023.10.1
tqdm>=4.66.1
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
duckdb 